Remove x86 from CI
The package may work well on x86. But testing would take some work. For example

bstring(128)

Prints a 64 bit string on x64, and 32 bit string on x86.

Testing in runtests.jl could be modified. But it would be a pain.
We also want jldocstring. How do we handle that. I would not want to junk it
up with UInt32 and UInt64 everywhere.